For #269: Move Diff callbacks into other files
