+++ app/src/main/java/com/google/samples/apps/sunflower/adapters/GardenPlantingAdapter.kt
@@ -22,6 +22,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.findNavController
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.google.samples.apps.sunflower.GardenFragmentDirections
@@ -74,5 +75,22 @@ class GardenPlantingAdapter(
                 executePendingBindings()
             }
         }
+    }
+}
+
+private class GardenPlantDiffCallback : DiffUtil.ItemCallback<PlantAndGardenPlantings>() {
+
+    override fun areItemsTheSame(
+        oldItem: PlantAndGardenPlantings,
+        newItem: PlantAndGardenPlantings
+    ): Boolean {
+        return oldItem.plant.plantId == newItem.plant.plantId
+    }
+
+    override fun areContentsTheSame(
+        oldItem: PlantAndGardenPlantings,
+        newItem: PlantAndGardenPlantings
+    ): Boolean {
+        return oldItem.plant == newItem.plant
     }
 }
